refactor(identity-auth): only store raw tokens and nonce in storage, rename `verifyToken` to `verifyTokens`, add `decodeTokens` method
We now only store the raw access token string under `gu.access_token`, and the raw id token string + nonce user
`gu.id_token` rather than the entire decoded token. This is because it was possible to modify what was in local
storage (e.g. a token's claims), verify the tokens, and use the modified values in your application, which can be
a security issue. By only storing the minimum information in local storage we can eliminate this risk, and instead
generate what is required when the tokens are decoded from local storage.

To fascilitate this we added a public `decodeTokens` method to the `Token` class.

We also rename `verifyToken` to `verifyTokens` as that better describes what is happening, i.e verifying both the access
and id tokens.