fix(identity-auth): update oauth timeout to 30 seconds
Updates the default oauth timeout from 12 seconds to 30 seconds as
we were seeing a few sentry issues which had timeout errors.
Hopefully by increasing the timeout the number of issues will
reduce.

An `oauthTimeout` parameter was also added to the options, just in
case this needs to be customised on a per application level.