create real make targets for project tasks
the wildcard target (`%`) causes it’s prerequisites to run before *all* `.PHONY` tasks, which is not what we want. it was a hack, but adding keeping a list up-to-date is hard by hand. we’re already doing it for the readme, so let’s also do it for the makefile.