react-crossword A11y improvements (#1966)
- Reduce the verbosity of the aria labels on the grid cells
- Remove the application role from the parent div as it is making the crossword not very visible to screen readers
- We have received feedback from a user that the labels are too verbose and make the crossword very difficult to use. The  information that was previously being read out can in fact be inferred by the position on the grid and the previous navigation context.

**previously** 
Every cell had information about what clue it was part of and how many letters there were in the word
Every Cell -> `Letter 2 of 4-across: Life is in a mess (5 letters). Also, letter 1 of 5-down Life is always in a mess (2 letters).`

**now** 
Only the cells at the start of a word have info about the clue and length. This gives the context for the surrounding letters. This should reduce the overload of information read out.
First Cell -> `4-across: Life is in a mess (5 letters).`